requirements.txt
openai-whisper
streamlit
gtts
sounddevice
scipy
scikit-learn
pandas
numpy
